refactor: moved jurisdiction screen and modal to cfd folder
