+++ packages/wallets/src/features/cfd/modals/index.ts
@@ -1,3 +1,4 @@
 export * from './DxtradeEnterPasswordModal';
+export * from './JurisdictionModal';
 export * from './MT5AccountTypeModal';
 export * from './MT5PasswordModal';
