+++ packages/wallets/src/features/cfd/screens/JurisdictionScreen/JurisdictionCard.scss
