+++ packages/wallets/src/features/cfd/screens/JurisdictionScreen/JurisdictionCardRow.scss
