+++ packages/wallets/src/features/cfd/screens/JurisdictionScreen/JurisdictionCardTag.scss
