+++ packages/wallets/src/features/cfd/screens/JurisdictionScreen/JurisdictionScreen.scss
@@ -1,4 +1,4 @@
-.wallets-jurisdiction-modal {
+.wallets-jurisdiction-screen {
     align-items: center;
     border-radius: 16px;
     display: flex;
