+++ packages/wallets/src/features/cfd/screens/JurisdictionScreen/index.ts
@@ -1,0 +1,1 @@
+export { default as JurisdictionScreen } from './JurisdictionScreen';
