+++ packages/wallets/src/features/cfd/screens/JurisdictionScreen/jurisdiction-contents/jurisdiction-bvi-contents.ts
