+++ packages/wallets/src/features/cfd/screens/JurisdictionScreen/jurisdiction-contents/jurisdiction-svg-contents.ts
