+++ packages/wallets/src/features/cfd/screens/JurisdictionScreen/jurisdiction-contents/jurisdiction-vanuatu-contents.ts
