+++ packages/wallets/src/features/cfd/screens/JurisdictionScreen/jurisdiction-contents/jurisdiction-verification-contents.ts
