+++ packages/wallets/src/features/cfd/screens/JurisdictionScreen/jurisdiction-contents/jurisdiction_maltainvest_contents.ts
