+++ packages/wallets/src/features/cfd/screens/JurisdictionScreen/jurisdiction-contents/props.types.ts
